Factor out resolve-interactive-styles
